Support running unit tests on iOS device
Added a minimal Xcode project with a iOS Application target that
provides a minimal UIApplicationDelegate that can host / run the unit
tests.

The target is intentionally setup as an "Application"-type target rather
than an XCTest bundle as it is generally easier to run application
targets on device (vs iOS Simulator).

Explicitly enabled build settings:

  Release configuration: NDEBUG=1

The commited shared .xcscheme is set to build and tests in Release.